src/requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
python-dotenv
mistralai
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain-google-genai
langgraph-cli